Initial Alstrup Macro Add Leaf Build (NOT WORKING)
Attempted to follow the Alstrup paper and get constant time leaf additions, but could not get it to work before the end of the semester. The paper does not give enough detail on how to maintain macro nodes within a tree induced by taking all the macro nodes from the original tree, and maintain the rank of each node in the tree in constant time.